Create a writable store for allScenarios
